+++ app/src/main/java/com/dede/android_eggs/util/ContextExt.kt
@@ -2,10 +2,13 @@ package com.dede.android_eggs.util
 
 import android.annotation.SuppressLint
 import android.app.Activity
+import android.content.ClipData
+import android.content.ClipboardManager
 import android.content.Context
 import android.content.Intent
 import android.widget.Toast
 import androidx.annotation.StringRes
+import androidx.core.content.getSystemService
 import com.dede.android_eggs.R
 import com.google.android.material.internal.ContextUtils
 
@@ -27,4 +30,10 @@ fun Context.toast(text: CharSequence, duration: Int = Toast.LENGTH_SHORT) {
 fun Context.createChooser(target: Intent): Intent {
     return Intent.createChooser(target, getString(R.string.title_open_with))
         .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+}
+
+fun Context.copy(text: String) {
+    val service = getSystemService<ClipboardManager>() ?: return
+    service.setPrimaryClip(ClipData.newPlainText(null, text))
+    toast(android.R.string.copy)
 }